lol commit

+++ cycle/sieveErastosthenes.cpp
@@ -7,14 +7,15 @@ using namespace std;
 int main(){
     int n,ct=0;
     cin>>n;
-    vector<int> a(n);
-    for(int i=0;i<n;i++)
+    vector <long> a(n);
+    for(long i=0;i<n;i++)
         a[i]=i;
-    for(int p=2;p<n;p++){
+    for(long p=2;p<n;p++){
         if(a[p]!=0){
             ct++;
             cout<<ct<<" "<<a[p]<<endl;
-            for(int j =p*p;j<n;j+=p)
+
+            for(unsigned long j =p*p;j<n;j+=p)
                 a[j]=0;
         }
     }
